refactor: extract nvidia-smi interaction into its own package
Move everything that talks to nvidia-smi out of the exporter into a dedicated internal package: query execution, CSV parsing, field discovery, and raw value interpretation. The query call is renamed from scrape to query, since scrape already has a Prometheus meaning. Field resolution now returns the query fields together with their returned names and the identity fields that label the gpu_info metric.

No behavior change, this is a pure relocation to prepare for decoupling metric collection from the scrape path. Part of #405.
