+++ internal/nvidiasmi/testdata/fields.txt
